Add modern Modeller headers & identifiers
Add headers as if we ran this modeling using the
latest version of Modeller. This identifies the
template and alignment used in the output models.

Use the correct residue numbering in the
Modeller run so we don't have to renumber the models
after the fact.

Use the final deposited name of the NMR structure (5TVZ).