Add combined alignment to generate oneshot model
This model was actually generated in two steps;
the second step uses the model from the first step
as one of the templates. To simplify the production of
an mmCIF file, provide a single alignment that maps
a crystal structure template to each region of the final
model. (While this could in principle be used to generate
the final model in one step, in practice it is unlikely
that sufficient sampling could be done.) Add modern headers
corresponding to this alignment to the final model; these
will be picked up by IMP.pmi.mmcif and stored in the mmCIF
file.